Cleanup codebase (#668)
* Rename projects

* Remove deprecated code

* Removes deprecated code

* Add license header

* Remove unnecessary usings

* Clean up tests

* Remove SampleAdministrators configuration from the DI container

* Defensive programming

* Refactor code to method to avoid repetition and clean up code

* Add an exception handling middleware

* Add an extension method to expose the middleware through IApplicationBuilder

* Define a model for the response data

* Add middleware to the request pipeline

* Remove repetitive try..catch{} blocks

* Remove unnecessary catch blocks

* Remove deprecated file

* Remove unassigned trace dictionary and repeated code block

* Add license header and remove try..catch block

* Add defensive programming
