+++ GraphExplorerPermissionsService/PermissionsService.csproj
