+++ GraphWebApi/Controllers/KnownIssuesController.cs
@@ -6,9 +6,7 @@ using KnownIssuesService.Interfaces;
 using KnownIssuesService.Models;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.DataContracts;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using UtilityService;
@@ -25,8 +23,9 @@ namespace GraphWebApi.Controllers
 
         public KnownIssuesController(IKnownIssuesService knownIssuesService, TelemetryClient telemetryClient)
         {
-            _telemetryClient = telemetryClient;
+            UtilityFunctions.CheckArgumentNull(telemetryClient, nameof(telemetryClient));
             UtilityFunctions.CheckArgumentNull(knownIssuesService, nameof(knownIssuesService));
+            _telemetryClient = telemetryClient;
             _knownIssuesService = knownIssuesService;
         }
 
@@ -36,20 +35,12 @@ namespace GraphWebApi.Controllers
         [HttpGet]
         public async Task<IActionResult> GetKnownIssues()
         {
-            try
-            {
-                _telemetryClient?.TrackTrace("Request to query the list of known issues",
-                                             SeverityLevel.Information,
-                                             _knownIssuesTraceProperties);
+            _telemetryClient?.TrackTrace("Request to query the list of known issues",
+                                            SeverityLevel.Information,
+                                            _knownIssuesTraceProperties);
 
-                List<KnownIssue> result = await _knownIssuesService.QueryBugsAsync();
-                return result == null ? NotFound() : Ok(result);
-            }
-            catch(Exception ex)
-            {
-                _telemetryClient?.TrackException(ex, _knownIssuesTraceProperties);
-                return new JsonResult(ex.Message) { StatusCode = StatusCodes.Status500InternalServerError };
-            }
+            List<KnownIssue> result = await _knownIssuesService.QueryBugsAsync();
+            return result == null ? NotFound() : Ok(result);
         }
     }
 }
