+++ GraphWebApi/GraphWebApi.csproj
@@ -61,8 +61,8 @@
     <ProjectReference Include="..\ChangesService\ChangesService.csproj" />
     <ProjectReference Include="..\CodeSnippetsReflection\CodeSnippetsReflection.csproj" />
     <ProjectReference Include="..\FileService\FileService.csproj" />
-    <ProjectReference Include="..\GraphExplorerPermissionsService\GraphExplorerPermissionsService.csproj" />
-    <ProjectReference Include="..\GraphExplorerSamplesService\GraphExplorerSamplesService.csproj" />
+    <ProjectReference Include="..\GraphExplorerPermissionsService\PermissionsService.csproj" />
+    <ProjectReference Include="..\GraphExplorerSamplesService\SamplesService.csproj" />
     <ProjectReference Include="..\KnownIssuesService\KnownIssuesService.csproj" />
     <ProjectReference Include="..\OpenAPIService\OpenAPIService.csproj" />
     <ProjectReference Include="..\TelemetryService\TelemetrySanitizerService.csproj" />
@@ -70,9 +70,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Content Update="SampleQueriesPolicies\CategoryPolicies.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
     <Content Update="wwwroot\swagger\favicon-16x16.png">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
