+++ GraphWebApi/Middleware/GlobalExceptionMiddleware.cs
@@ -1,0 +1,16 @@
+﻿// ------------------------------------------------------------------------------------------------------------------------------------------------------
+//  Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
+// ------------------------------------------------------------------------------------------------------------------------------------------------------
+
+using Microsoft.AspNetCore.Builder;
+
+namespace GraphWebApi.Middleware
+{
+    public static class GlobalExceptionMiddleware
+    {
+        public static void UseGlobalExceptionMiddleware(this IApplicationBuilder app)
+        {
+            app.UseMiddleware<ExceptionMiddleware>();
+        }
+    }
+}
