--- GraphWebApi/Models/SamplesAdministrators.cs
@@ -1,10 +1,0 @@
-﻿namespace GraphWebApi.Models
-{
-    /// <summary>
-    /// Defines a list of the samples administrators 
-    /// </summary>
-    public class SamplesAdministrators
-    {
-        public string[] Administrators { get; set; }
-    }
-}
