+++ GraphWebApi/Startup.cs
@@ -28,6 +28,7 @@ using TelemetrySanitizerService;
 using OpenAPIService.Interfaces;
 using OpenAPIService;
 using KnownIssuesService.Interfaces;
+using GraphWebApi.Middleware;
 
 namespace GraphWebApi
 {
@@ -95,7 +96,6 @@ namespace GraphWebApi
             services.AddSingleton<IKnownIssuesService, KnownIssuesService.Services.KnownIssuesService>();
             services.AddHttpClient<IHttpClientUtility, HttpClientUtility>();
             services.AddControllers().AddNewtonsoftJson();
-            services.Configure<SamplesAdministrators>(Configuration);
 
             // Localization
             services.Configure<RequestLocalizationOptions>(options =>
@@ -127,6 +127,7 @@ namespace GraphWebApi
             {
                 app.UseHsts();
             }
+            app.UseGlobalExceptionMiddleware();
             app.UseStaticFiles(new StaticFileOptions
             {
                 DefaultContentType = "text/plain",
