+++ PermissionsService.Test/PermissionsService.Test.csproj
@@ -62,7 +62,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\GraphExplorerPermissionsService\GraphExplorerPermissionsService.csproj" />
+    <ProjectReference Include="..\GraphExplorerPermissionsService\PermissionsService.csproj" />
     <ProjectReference Include="..\MockTestUtility\MockTestUtility.csproj" />
   </ItemGroup>
