+++ SamplesService.Test/SamplesServiceShould.cs
@@ -281,721 +281,5 @@ namespace SamplesService.Test
         }
 
         #endregion
-
-        #region Update Sample Queries List Tests
-
-        [Fact]
-        public void UpdateSampleQueryInListOfSampleQueries()
-        {
-            /* Arrange */
-
-            // Create a list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false
-                }
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            /* Create a 'Users' sample query model object that will update an existing 'Users'
-             * sample query object in the list of sample queries.
-             */
-            SampleQueryModel sampleQueryModel = new SampleQueryModel()
-            {
-                Category = "Users",
-                Method = SampleQueryModel.HttpMethods.GET,
-                HumanName = "my direct reports",
-                RequestUrl = "/v1.0/me/directReports",
-                DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                SkipTest = false,
-                Tip = "This item has been updated." // update field
-            };
-
-            Guid id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b");
-
-            /* Act */
-
-            // Update the provided sample query model into the list of sample queries
-            SampleQueriesList updatedSampleQueriesList = GraphExplorerSamplesService.Services.SamplesService.UpdateSampleQueriesList
-                (sampleQueriesList, sampleQueryModel, id);
-
-            // Assert - item on index[1] has an updated property value
-            Assert.Equal(sampleQueryModel.Tip, updatedSampleQueriesList.SampleQueries[1].Tip);
-        }
-
-        [Fact]
-        public void ThrowInvalidOperationExceptionIfUpdateSampleQueriesListSampleQueryIdNotFoundInCollection()
-        {
-            /* Arrange */
-
-            // Create a list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false
-                }
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            /* Create a 'Users' sample query model object that should update an existing 'Users'
-             * sample query object in the list of sample queries.
-             */
-            SampleQueryModel sampleQueryModel = new SampleQueryModel()
-            {
-                Category = "Users",
-                Method = SampleQueryModel.HttpMethods.GET,
-                HumanName = "my direct reports",
-                RequestUrl = "/v1.0/me/directReports",
-                DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                SkipTest = false,
-                Tip = "This item has been updated." // updated field
-            };
-
-            Guid id = Guid.Parse("5484add6-d4be-4560-82ef-31ef738775e8"); // non-existent id
-
-            /* Act and Assert */
-
-            // Attempt to update the provided sample query model into the list of sample queries with a non existent id
-            Assert.Throws<InvalidOperationException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.UpdateSampleQueriesList
-                (sampleQueriesList, sampleQueryModel, id));
-        }
-
-        [Fact]
-        public void ThrowArgumentNullExceptionIfUpdateSampleQueriesListSampleQueryModelParameterIsNull()
-        {
-            /* Arrange */
-
-            // Create list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false
-                }
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            SampleQueryModel sampleQueryModel = null;
-
-            Guid id = Guid.NewGuid();
-
-            /* Act and Assert */
-
-            // Null sample query model
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.UpdateSampleQueriesList(sampleQueriesList, sampleQueryModel, id));
-        }
-
-        [Fact]
-        public void ThrowArgumentNullExceptionIfUpdateSampleQueriesListSampleQueryIdParameterIsEmpty()
-        {
-            /* Arrange */
-
-            // Create list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                }
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            SampleQueryModel sampleQueryModel = new SampleQueryModel()
-            {
-                Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                Category = "Groups",
-                Method = SampleQueryModel.HttpMethods.GET,
-                HumanName = "all groups in my organization",
-                RequestUrl = "/v1.0/groups",
-                DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                SkipTest = false
-            };
-
-            Guid id = Guid.Empty;
-
-            /* Act and Assert */
-
-            // Empty sample query id
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.UpdateSampleQueriesList(sampleQueriesList, sampleQueryModel, id));
-        }
-
-        [Fact]
-        public void ThrowArgumentNullExceptionIfUpdateSampleQueriesListSampleQueriesListParameterIsNullOrEmptyCollection()
-        {
-            /* Arrange */
-
-            SampleQueriesList nullSampleQueriesList = null;
-
-            SampleQueriesList emptySampleQueriesList = new SampleQueriesList();
-
-            SampleQueryModel sampleQueryModel = new SampleQueryModel()
-            {
-                Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                Category = "Getting Started",
-                Method = SampleQueryModel.HttpMethods.GET,
-                HumanName = "my manager",
-                RequestUrl = "/v1.0/me/manager",
-                DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                SkipTest = false
-            };
-
-            Guid id = Guid.NewGuid();
-
-            /* Act and Assert */
-
-            // Null sample queries list
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.UpdateSampleQueriesList(nullSampleQueriesList, sampleQueryModel, id));
-
-            // Empty sample queries list
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.UpdateSampleQueriesList(emptySampleQueriesList, sampleQueryModel, id));
-        }
-
-        #endregion
-
-        #region Add To Sample Queries List Tests
-
-        [Fact]
-        public void AddSampleQueryIntoEmptyListOfSampleQueries()
-        {
-            /* Arrange */
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList();
-
-            SampleQueryModel sampleQueryModel = new SampleQueryModel()
-            {
-                Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                Category = "Getting Started",
-                Method = SampleQueryModel.HttpMethods.GET,
-                HumanName = "my manager",
-                RequestUrl = "/v1.0/me/manager",
-                DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                SkipTest = false
-            };
-
-            /* Act */
-
-            // Add the new sample query to the empty list of sample queries
-            sampleQueriesList = GraphExplorerSamplesService.Services.SamplesService.AddToSampleQueriesList
-                (sampleQueriesList, sampleQueryModel);
-
-            // Assert
-            Assert.NotEmpty(sampleQueriesList.SampleQueries);
-        }
-
-        [Fact]
-        public void AddSampleQueryIntoListOfSampleQueriesInHierarchicalOrderOfCategories()
-        {
-            /* Arrange */
-
-            // Create a list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false},
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            /* Create a new 'Users' sample query model object that will be inserted into
-             * the list of sample queries in a hierarchical order of categories.
-             */
-            SampleQueryModel sampleQueryModel = new SampleQueryModel()
-            {
-                Category = "Users",
-                Method = SampleQueryModel.HttpMethods.GET,
-                HumanName = "all users in the organization",
-                RequestUrl = "/v1.0/users",
-                DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/users",
-                SkipTest = false,
-            };
-
-            /* Act */
-
-            // Assign a new Id to the new sample query
-            sampleQueryModel.Id = Guid.NewGuid();
-
-            // Add the new sample query to the list of sample queries
-            SampleQueriesList newSampleQueriesList = GraphExplorerSamplesService.Services.SamplesService.AddToSampleQueriesList
-                (sampleQueriesList, sampleQueryModel);
-
-            // Assert - the new sample query should be inserted at index[2] of the list of sample queries
-            Assert.Equal(sampleQueryModel.Id, newSampleQueriesList.SampleQueries[2].Id);
-        }
-
-        [Fact]
-        public void AddSampleQueryToTopOfListOfSampleQueriesIfHighestRankedSampleQueryCategory()
-        {
-            /* Arrange */
-
-            // Create a list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false
-                },
-                 new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Outlook Mail",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my high important mail",
-                    RequestUrl = "/v1.0/me/messages?$filter=importance eq 'high'",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_messages",
-                    SkipTest = false
-                }
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            /* Create a new 'Users' sample query model object that will be inserted into
-             * the list of sample queries as the highest ranked sample category
-             */
-            SampleQueryModel sampleQueryModel = new SampleQueryModel()
-            {
-                Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                Category = "Users",
-                Method = SampleQueryModel.HttpMethods.GET,
-                HumanName = "my manager",
-                RequestUrl = "/v1.0/me/manager",
-                DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                SkipTest = false
-            };
-
-            /* Act */
-
-            // Assign a new Id to the new sample query
-            sampleQueryModel.Id = Guid.NewGuid();
-
-            // Add the new sample query to the list of sample queries
-            SampleQueriesList newSampleQueriesList = GraphExplorerSamplesService.Services.SamplesService.AddToSampleQueriesList
-                (sampleQueriesList, sampleQueryModel);
-
-            // Assert - the new sample query should be inserted at index[0] of the list of sample queries
-            Assert.Equal(sampleQueryModel.Id, newSampleQueriesList.SampleQueries[0].Id);
-        }
-
-        [Fact]
-        public void ThrowArgumentNullExceptionIfAddToSampleQueriesListSampleQueriesListParameterIsNull()
-        {
-            /* Arrange */
-
-            SampleQueriesList nullSampleQueriesList = null;
-
-            SampleQueryModel sampleQueryModel = new SampleQueryModel()
-            {
-                Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                Category = "Getting Started",
-                Method = SampleQueryModel.HttpMethods.GET,
-                HumanName = "my manager",
-                RequestUrl = "/v1.0/me/manager",
-                DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                SkipTest = false
-            };
-
-            /* Act and Assert */
-
-            // Assign a new Id to the new sample query
-            sampleQueryModel.Id = Guid.NewGuid();
-
-            // Null sample queries list
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.AddToSampleQueriesList(nullSampleQueriesList, sampleQueryModel));
-        }
-
-        [Fact]
-        public void ThrowArgumentNullExceptionIfAddToSampleQueriesListSampleQueryModelParameterIsNull()
-        {
-            /* Arrange */
-
-            // Create list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false
-                }
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            SampleQueryModel sampleQueryModel = null;
-
-            /* Act and Assert */
-
-            // Null sample query model
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.AddToSampleQueriesList(sampleQueriesList, sampleQueryModel));
-        }
-
-        #endregion
-
-        #region Remove Sample Query Tests
-
-        [Fact]
-        public void RemoveSampleQueryFromListOfSampleQueries()
-        {
-            /* Arrange */
-
-            // Create a list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false},
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            Guid idToDelete = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"); // id of sample query to delete from list of sample queries
-
-            /* Act */
-
-            // Remove the User's query from the list of sample queries
-            SampleQueriesList updatedSampleQueriesList = GraphExplorerSamplesService.Services.SamplesService.RemoveSampleQuery
-                (sampleQueriesList, idToDelete);
-
-            // Assert - reference to the deleted User's sample query with the given id should be null
-            Assert.Null(updatedSampleQueriesList.SampleQueries.Find(x => x.Id == idToDelete));
-        }
-
-        [Fact]
-        public void ThrowInvalidOperationExceptionIfRemoveSampleQuerySampleQueryIdNotFoundInCollection()
-        {
-            /* Arrange */
-
-            // Create a list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false
-                }
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            Guid id = Guid.Parse("e6819ecc-f5aa-4792-ac86-c25234383513"); // Non-existent id
-
-            /* Act and Assert */
-
-            // Attempt to remove a sample query from the list of sample queries with a non-existent id
-            Assert.Throws<InvalidOperationException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.RemoveSampleQuery
-                (sampleQueriesList, id));
-        }
-
-        [Fact]
-        public void ThrowArgumentNullExceptionIfRemoveSampleQuerySampleQueriesListParameterIsNullOrEmptyCollection()
-        {
-            /* Arrange */
-
-            SampleQueriesList nullSampleQueriesList = null;
-
-            SampleQueriesList emptySampleQueriesList = new SampleQueriesList();
-
-            Guid id = Guid.Parse("9fd2bcca-6597-4fd8-a6de-0119a0b94e20");
-
-            /* Act and Assert */
-
-            // Null sample queries list
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.RemoveSampleQuery(nullSampleQueriesList, id));
-
-            // Empty sample queries list
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.RemoveSampleQuery(emptySampleQueriesList, id));
-        }
-
-        [Fact]
-        public void ThrowArgumentNullExceptionIfRemoveSampleQuerySampleQueryIdParameterIsEmpty()
-        {
-            /* Arrange */
-
-            // Create list of sample queries
-            List<SampleQueryModel> sampleQueries = new List<SampleQueryModel>()
-            {
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("3482cc10-f2be-40fc-bcdb-d3ac35f3e4c3"),
-                    Category = "Getting Started",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my manager", RequestUrl = "/v1.0/me/manager",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_manager",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("48b62369-3974-4783-a5c6-ae4ea2d8ae1b"),
-                    Category = "Users",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "my direct reports",
-                    RequestUrl = "/v1.0/me/directReports",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/api/user_list_directreports",
-                    SkipTest = false
-                },
-                new SampleQueryModel()
-                {
-                    Id = Guid.Parse("7d5bac53-2e16-4162-b78f-7da13e77da1b"),
-                    Category = "Groups",
-                    Method = SampleQueryModel.HttpMethods.GET,
-                    HumanName = "all groups in my organization",
-                    RequestUrl = "/v1.0/groups",
-                    DocLink = "https://developer.microsoft.com/en-us/graph/docs/api-reference/v1.0/resources/group",
-                    SkipTest = false
-                }
-            };
-
-            SampleQueriesList sampleQueriesList = new SampleQueriesList()
-            {
-                SampleQueries = sampleQueries
-            };
-
-            Guid id = Guid.Empty;
-
-            /* Act and Assert */
-
-            // Empty sample query Id
-            Assert.Throws<ArgumentNullException>(() =>
-                GraphExplorerSamplesService.Services.SamplesService.RemoveSampleQuery(sampleQueriesList, id));
-        }
-
-        #endregion
     }
 }
